requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Document Parsers
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
pdfplumber==0.11.0
pillow==10.2.0

# Protocol Adapters
smbprotocol==1.12.0
requests==2.31.0
httpx==0.26.0
gitpython==3.1.41
boto3==1.34.34
azure-storage-blob==12.19.0
google-cloud-storage==2.14.0

# SharePoint
office365-rest-python-client==2.5.3

# Security & Auth
cryptography==42.0.2
python-jose[cryptography]==3.3.0
hvac==2.1.0

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
aiofiles==23.2.1
pyyaml==6.0.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.6
pytest-cov==4.1.0
mypy==1.8.0
black==24.1.1
ruff==0.2.0
bandit==1.7.6
safety==3.2.0
